# Vectorized MemoryExtensions.CommonPrefixLength