@stephentoub 2022-04-20
Thanks.  This LGTM, but @tannergooding should review as well.

> Benchmarks aren't in https://github.com/dotnet/performance. Should these be added? Or is it covered via usage in other parts, or too niche to have it covered?

It's too new; last I checked, the performance repo hadn't ingested a runtime with these APIs.  If it now has them, it'd be worth adding a few.